Break typet into multiple modules.
