+++ setup.cfg
@@ -32,8 +32,3 @@ addopts = -vvra
 
 [mypy]
 ignore_missing_imports=True
-
-[pylint]
-disable=no-name-in-module,invalid-metaclass,unsupported-membership-test,no-value-for-parameter,protected-access,function-redefined,too-few-public-methods,invalid-sequence-index,bad-continuation,import-error,invalid-name,no-member,locally-disabled,locally-enabled,redefined-outer-name,redefined-variable-type
-reports=no
-known-standard-library=typing
