+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 skipsdist = True
-envlist = py{27,33,34,35,36}
+envlist = py{27,34,35,36}
 
 [testenv]
 setenv =
@@ -8,16 +8,18 @@ setenv =
 usedevelop = True
 passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
 deps =
-    pytest >= 3.2
+    pytest >= 3.2, < 4
     pytest-cov
     coveralls
     pep257
     flake8
     pylint >= 1.7
+    py{34,35,36}: mypy >= 0.501
 commands =
-    py{27,33,34,35}: py.test --cov=typet --basetemp={envtmpdir} --ignore=tests/py36 {posargs}
+    py{27,34,35}: py.test --cov=typet --basetemp={envtmpdir} --ignore=tests/py36 {posargs}
     py36: py.test --cov=typet --basetemp={envtmpdir} {posargs}
     - coveralls
     pep257 setup.py typet tests
     flake8 setup.py typet
-    pylint --rcfile=setup.cfg setup.py typet
+    pylint setup.py typet
+    py{34,35,36}: - mypy --config=setup.cfg setup.py typet
